% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{all_sites}
\alias{all_sites}
\title{Available USGS sites data}
\format{
A data frame of 301 rows and one column representing the site
number of a particular site
\describe{
\item{site_no}{The site number}
}
}
\usage{
all_sites
}
\description{
All the sites that are provided by the USGS. This dataset is
a subset of the rbi water year all sites dataset
}
\keyword{datasets}
