% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gagesII.R
\name{basin_id}
\alias{basin_id}
\title{Retrieve site classification (Reference and Non-reference sites) from USGS GAGES II}
\usage{
basin_id(site_num)
}
\arguments{
\item{site_num}{a "character" vector that contains the study site numbers.}
}
\value{
returns a data frame with 10 variables which includes the drainage area and site classification
}
\description{
Retrieve site classification (Reference and Non-reference sites) from USGS GAGES II
}
