% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gagesII.R
\name{dam_removal}
\alias{dam_removal}
\title{Retrieve dam removal data for specified sites directly from the USGS Gages II website. This time series data recorded between 1931-01-01 and 2014-12-31
dam_removal}
\usage{
dam_removal(site_num)
}
\arguments{
\item{site_num}{a "character" vector that contains the study site numbers.}
}
\value{
returns a data frame with 7 variables which includes the year when the dam was removed and the site location.
}
\description{
Retrieve dam removal data for specified sites directly from the USGS Gages II website. This time series data recorded between 1931-01-01 and 2014-12-31
dam_removal
}
\examples{
# returns a dataframe with the dam removals for specified sites.
library(easyrbi)
dam_removal("01034500")
}
