% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gagesII.R
\name{dam_trends}
\alias{dam_trends}
\title{Create a dataframe containing dam removal and hydrology data from the USGS and USGS GAGES II websites.}
\usage{
dam_trends(site_num, startDate, endDate)
}
\arguments{
\item{site_num}{a "character" vector that contains the study site numbers.}

\item{startDate}{a "date" following the Y-m-d format.}

\item{endDate}{a "date" following the Y-m-d format.}
}
\value{
returns a dataframe with 21 variables which include the dams removed and corresponding summary statistics from the sens.Slope and Kendall analysis.
}
\description{
Create a dataframe containing dam removal and hydrology data from the USGS and USGS GAGES II websites.
}
\examples{
# returns RBI summary statistics and dam removal data for given sites over the specified period
library(easyrbi)
dam_trends("01567000", "1970-10-01", "1990-09-30")

}
