% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{rbi}
\alias{rbi}
\title{Calculate the Richards Baker Flashiness Index (RBI) for given gauge stations}
\usage{
rbi(q)
}
\arguments{
\item{q}{a dataframe containing the daily discharge data.}
}
\description{
Calculate the Richards Baker Flashiness Index (RBI) for given gauge stations
}
