% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{rbi_df}
\alias{rbi_df}
\title{Create a dataframe with the calculated annual Richards Baker Flashiness Index (RBI) for specified sites and duration}
\usage{
rbi_df(site_num, startDate, endDate)
}
\arguments{
\item{site_num}{a "character" vector that contains the study site numbers.}

\item{startDate}{a "date" following the Y-m-d format.}

\item{endDate}{a "date" following the Y-m-d format.}
}
\value{
returns a data.frame with the waterYear variable and RBI values for specified sites saved in columns.
}
\description{
Create a dataframe with the calculated annual Richards Baker Flashiness Index (RBI) for specified sites and duration
}
\examples{
# return a dataframe with annual RBI values for specified USGS gauge sites over specified duration
library(easyrbi)
rbi_df(c("01564500", "01567000"), "1970-10-01", "1980-09-30")
}
