% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{sitedata}
\alias{sitedata}
\title{Retrieve hydrology data from the USGS website for given sites and duration}
\usage{
sitedata(site_num, startDate, endDate)
}
\arguments{
\item{site_num}{a "character" vector that contains the study site numbers.}

\item{startDate}{a "date" following the Y-m-d format.}

\item{endDate}{a "date" following the Y-m-d format.}
}
\value{
returns a data.frame with 18 variables.
}
\description{
Retrieve hydrology data from the USGS website for given sites and duration
}
\examples{
# return stream hydrology data for given USGS sites
library(easyrbi)
sitedata(c("01564500", "01567000"), "1970-10-01", "1980-09-30")
}
