% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumstats.R
\name{trendAnalysis}
\alias{trendAnalysis}
\title{Calculate the sens.slope and MannKendall values
trendAnalysis}
\usage{
trendAnalysis(x)
}
\arguments{
\item{x}{a list of type spec_tbl_df/tbl_df/tbl/data.frame that has one column with consecutive times of class 'numeric'. The time variable can also be a 'Date' class.}
}
\description{
Calculate the sens.slope and MannKendall values
trendAnalysis
}
\examples{

df <- data.frame (year  = seq(from = 2010, to = 2020, by = 1),
site_num = seq(from = 1, to= 2, by = .1))

trendAnalysis(df$site_num)


}
