% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumstats.R
\name{trends}
\alias{trends}
\title{Returns the results from trendAnalysis as a dataframe
trends}
\usage{
trends(site = all_sites$site_no)
}
\arguments{
\item{site}{a "character" vector that contains the study site numbers. The default for the site argument is character vector with 301 study site names/numbers.}
}
\value{
returns a list/data.frame with 15 variables calculated using trendAnalysis. The data.frame has 3 columns that are character vectors and 12 that are numeric.
}
\description{
Returns the results from trendAnalysis as a dataframe
trends
}
\examples{
library(easyrbi)

# return summary statistics for all USGS sites given in the rbiWy_dfAll dataframe
trends()

# This function allows the user to input the study site name/number
trends(site = c("01011000", "01042500"))
}
