% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gagesII.R
\name{dam_removal}
\alias{dam_removal}
\title{Retrieve dam removal time series data from the USGS Gages II website.}
\usage{
dam_removal(site_num)
}
\arguments{
\item{site_num}{a "character" vector that contains the study site numbers.}
}
\value{
returns a data frame with 7 variables which includes the year when the dam was removed and the site location.
}
\description{
Retrieve dam removal time series data from the USGS Gages II website.
}
\examples{
# returns a dataframe with the dam removals for specified sites.
library(easyrbi)
study_sites <- c("01564500", "01567000", "01567500", "01568000",
"01570000", "01571500", "01573000", "01574000", "01574500", "01576500")
dam_removal(study_sites)
}
