% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gagesII.R
\name{dam_trends}
\alias{dam_trends}
\title{Retrieve dam removal and hydrology data from the USGS and USGS Gages II websites for given sites over a specified period. The dataframe contains calculated RBI summary statistics and dam removal time series data recorded between 1931-01-01 and 2014-12-31.
dam_trends}
\usage{
dam_trends(site_num, startDate, endDate)
}
\arguments{
\item{site_num}{a "character" vector that contains the study site numbers.}

\item{startDate}{a "date" following the Y-m-d format.}

\item{endDate}{a "date" following the Y-m-d format.}
}
\value{
returns a dataframe with 21 variables which include the dams removed and corresponding summary statistics from the sens.Slope and Kendall analysis.
}
\description{
Retrieve dam removal and hydrology data from the USGS and USGS Gages II websites for given sites over a specified period. The dataframe contains calculated RBI summary statistics and dam removal time series data recorded between 1931-01-01 and 2014-12-31.
dam_trends
}
