% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{rbi}
\alias{rbi}
\title{Calculate the Richard Baker Index (RBI) for given sites
rbi}
\usage{
rbi(q)
}
\arguments{
\item{q}{a dataframe containing the daily discharge data.}
}
\description{
Calculate the Richard Baker Index (RBI) for given sites
rbi
}
