% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{rbiWy_dfAll}
\alias{rbiWy_dfAll}
\title{Calculated RBI data for 301 sites USGS sites data}
\format{
A data frame of 15050 rows and 3 columns.
\describe{
\item{waterYear}{The water year}
\item{site_no}{USGS gauge site number}
\item{rbi}{annual RBI values for the 301 study sites}
}
}
\usage{
rbiWy_dfAll
}
\description{
RBI values are calculated for 301 USGS sites and stored in the dataframe. This data frame contains water year and annual RBI value for each site. Each column
from 01011000 to 01576500 contains RBI values for each site.
}
\keyword{datasets}
