% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{rbiWy_dfAll}
\alias{rbiWy_dfAll}
\title{Calculated RBI values for an example dataset containing 301 USGS sites.}
\format{
A data frame of 15050 rows and 3 columns.
\describe{
\item{waterYear}{The water year}
\item{site_no}{USGS gauge site number}
\item{rbi}{annual RBI values for the 301 study sites}
}
}
\usage{
rbiWy_dfAll
}
\description{
The data frame contains the annual RBI values.
}
\keyword{datasets}
