% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumstats.R
\name{trends}
\alias{trends}
\title{Create a dataframe containing summary statistics from the MannKendall and sens.slope analysis for set of USGS gauge stations over given duration.
trends}
\usage{
trends(x)
}
\arguments{
\item{x}{a data.frame containing the waterYear and RBI values for sites in one column.}
}
\value{
returns a list/data.frame with 15 variables calculated using trendAnalysis. The data.frame has 3 columns that are character vectors and 12 that are numeric.
}
\description{
Create a dataframe containing summary statistics from the MannKendall and sens.slope analysis for set of USGS gauge stations over given duration.
trends
}
\examples{
library(easyrbi)
library(dplyr)

# return summary statistics for all USGS sites given in a dataframe containing annual RBI values
data <- rbi_df(c("01564500", "01567000"), "1970-10-01", "1980-09-30")
trends(x = data)
}
